requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
fastapi==0.115.0
fonttools==4.60.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
# httpx==0.27.2
idna==3.10
iniconfig==2.1.0
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.3.3
orjson==3.10.7
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.5
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.3
six==1.17.0
sniffio==1.3.1
starlette==0.38.6
supabase==2.4.6
typing_extensions==4.15.0
uvicorn==0.30.6
# uvloop==0.21.0
watchfiles==1.1.0
